Clarify which DNSSEC version we are referring to
The current text can be interpreted incorrectly. It essentially says "DNSSEC" is formally version 3 .... however, earlier versions of DNSSEC were ... significantly less visible ... Here DNSSEC means the version of the protocol initially defined in ..." 

The use of the word "initially" in the last sentence, in conjunction with some low-numbered RFC numbers, can lead the reader to think that you are referring to version 1. This PR aims to make it clearer that we are pointing to version 3.